(wip) prevent propose messages from being dropped
There was a race where a proposal could be dropped. If
the prop channel was full and the main run goroutine was
busy processing a message, a call to Propose would take
the default case. Now there is no default, and no race;
if the node hasn't been stopped, Propose always sends
the message.

Also, it's faster now for some reason. (TODO(kr): list
benchmark changes here and investigate why.)